feat: implement X and Y field selectors for line chart settings
